+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 ===================
 
+## 2.0.0-beta.14 - 2018-09-10
+
+### Added
+- Added missing Craft 3 upgrade migration.
+
+### Fixed
+- Fixed a bug where profile couldn’t be filled properly on user registration.
+- Fixed bug where too many arguments were passed to `\dukt\social\services\LoginProviders::getLoginProvider()` in `\dukt\social\controllers\LoginProvidersController::actionOauth()`.
+
 ## 2.0.0-beta.13 - 2018-08-29
 
 ### Fixed